+++ powershell/install-step.ps1
@@ -7,12 +7,11 @@ mkdir $install_dir >$null
 # Determine filename to download
 $response = Invoke-Webrequest -Uri "https://api.github.com/repos/smallstep/cli/releases/latest" -UseBasicParsing
 $release = $response | ConvertFrom-Json
-$step_version = $release.tag_name.Substring(1)
 
 # Download the binary and uninstall script
 echo ""
 $exe_file = join-path $install_dir "step.exe"
-curl.exe -L -o $exe_file "https://dl.step.sm/s3/cli/s3-windows-installer/step-windows-$step_version.exe"
+curl.exe -L -o $exe_file "https://dl.step.sm/s3/cli/s3-windows-installer/step_latest_windows.exe"
 $uninstall_file = join-path $install_dir "uninstall-step.ps1"
 curl.exe -L -o $uninstall_file https://dl.step.sm/s3/cli/s3-windows-installer/uninstall-step.ps1
 echo ""